requirements.txt
apiai==1.2.3
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
aylien-apiclient==0.7.0
beautifulsoup4==4.7.1
bleach==3.1.0
cachetools==3.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
clarifai==2.5.2
Click==7.0
colorama==0.4.1
configparser==3.5.0
coverage==4.5.2
crayons==0.1.2
cryptography==2.4.2
dialogflow==0.5.2
docutils==0.14
dropbox==9.3.0
Flask==1.0.2
future==0.17.1
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-storage==1.13.2
google-resumable-media==0.3.2
googleapis-common-protos==1.5.6
grpcio==1.18.0
html2text==2018.1.9
httplib2==0.12.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
lxml==4.3.0
MarkupSafe==1.1.0
matrix-client==0.3.2
mock==2.0.0
more-itertools==5.0.0
mypy==0.600
numpy==1.16.0
oauthlib==3.0.0
pbr==5.1.1
pkginfo==1.5.0.1
pluggy==0.8.1
prompt-toolkit==2.0.7
protobuf==3.6.1
py==1.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycodestyle==2.3.1
pycparser==2.19
Pygments==2.3.1
pyOpenSSL==18.0.0
PySocks==1.6.8
pytest==4.1.1
python-chess==0.22.0
pytz==2018.9
readme-renderer==24.0
requests==2.21.0
requests-oauthlib==1.2.0
requests-toolbelt==0.8.0
rsa==4.0
simple-salesforce==0.74.2
six==1.12.0
soupsieve==1.7.2
tqdm==4.29.1
tweepy==3.7.0
twine==1.12.1
typed-ast==1.1.2
typing==3.6.6
urllib3==1.24.1
watson-developer-cloud
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wit==5.1.0
zulip==0.5.7
zulip-bots==0.5.7
zulip-botserver==0.5.7
selenium
-e ./zulip
-e ./zulip_bots
-e ./zulip_botserver
